Renames routes so they are all consistent
