+++ server/routes/about-person/routes.ts
@@ -1,9 +1,9 @@
 import { Router } from 'express'
-import AboutPopController from './AboutPopController'
+import AboutPersonController from './AboutPersonController'
 import URLs from '../URLs'
 
 export default function setupAboutPopRoutes(router: Router) {
-  const controller = new AboutPopController()
+  const controller = new AboutPersonController()
 
   router.get(URLs.ABOUT_POP, controller.get)
 }
