+++ server/routes/index.ts
@@ -1,19 +1,19 @@
 import { RequestHandler, Router } from 'express'
 import asyncMiddleware from '../middleware/asyncMiddleware'
 import type { Services } from '../services'
-import setupCreateGoalRoutes from './createGoal/routes'
-import setupRemoveGoalRoutes from './removeGoal/routes'
-import setupChangeGoalRoutes from './changeGoal/routes'
-import setupAboutPopRoutes from './aboutPop/routes'
+import setupCreateGoalRoutes from './create-goal/routes'
+import setupRemoveGoalRoutes from './remove-goal/routes'
+import setupChangeGoalRoutes from './change-goal/routes'
+import setupAboutPopRoutes from './about-person/routes'
 import { Page } from '../services/auditService'
-import setupReferenceDataRoutes from './ReferenceData/routes'
-import setupPlanOverviewRoutes from './planOverview/routes'
+import setupReferenceDataRoutes from './reference-data/routes'
+import setupPlanOverviewRoutes from './plan-overview/routes'
 import setupAgreePlanRoutes from './agree-plan/routes'
 import URLs from './URLs'
 import setupAddStepsRoutes from './add-steps/routes'
-import setupAchieveGoalRoutes from './achieveGoal/routes'
+import setupAchieveGoalRoutes from './achieve-goal/routes'
 import setupUpdateGoalRoutes from './update-goal/routes'
-import setupViewGoalDetailsRoutes from './viewGoalDetails/routes'
+import setupViewGoalDetailsRoutes from './view-goal-details/routes'
 import setupPlanHistoryRoutes from './plan-history/routes'
 
 export default function routes(services: Services): Router {
