+++ server/routes/remove-goal/RemoveGoalController.ts
