+++ server/routes/remove-goal/locale-delete.json
