+++ server/routes/remove-goal/routes.ts
