+++ server/routes/view-goal-details/locale.json
