+++ server/routes/view-goal-details/routes.ts
